Add ability for multiple minima regions
